Client.js: Add ending character at the end of the message
Packets that are sent to the server must contain a null ending character
to indicate the end of the message.

Pique handles this already due to the way its designed and
SpadesX only very recently started handling this but as it is
a client (or in this case a library) issue this should be handled
here.